refactor: add separate SearchBar component
