+++ src/components/SearchBar.jsx
@@ -1,0 +1,24 @@
+import { useState } from "react";
+
+export default function SearchBar({ changeProcessedProducts, changeReset }) {
+  const [searchInput, setSearchInput] = useState("");
+
+  function handleChange(e) {
+    e.preventDefault();
+    setSearchInput(e.target.value);
+    if (searchInput.length > 0) {
+      changeProcessedProducts(searchInput);
+    }
+  }
+  return (
+    <>
+      <input
+        type="text"
+        value={searchInput}
+        placeholder="Search"
+        onChange={(e) => handleChange(e)}
+      />
+      <button onClick={changeReset}>Reset</button>
+    </>
+  );
+}
